allow multiple paragraphs in blockquotes
